Added robustness against DOM changes
SVG injection was not working well for several cases.
This happene in a jQuery-UI button.
jQuery-UI redefines the DOM element by wrapping it inside other DOM elements.  As a result "this.element" is no longer referencing a DOM element by the time the SVG is loaded.
In order to make svgInject robust against this, this modification adds a unique class to the original element and the replacement is performed by using this class as the specifier.